requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Async HTTP
aiohttp==3.9.1


# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.0

aiomysql==0.2.0

# Task Queue
celery==5.3.4
# Langchain & LLM
langchain==0.0.352
langchain-google-genai==0.0.6
langsmith==0.0.70


# Vector Store
chromadb==0.4.18

# PDF Generation
reportlab==4.0.7
fpdf2==2.7.6

# Authentication & Security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration & Environment
pydantic==2.5.1
pydantic-settings==2.1.0
python-dotenv==1.0.0
pyyaml==6.0.1

# Monitoring & Observability
sentry-sdk[fastapi]==1.38.0
tenacity==8.2.3
aiobreaker==1.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

pre-commit==3.6.0

# Utilities
python-multipart==0.0.6
email-validator==2.1.0

